host/requirements.txt
pyserial
orgasm-package 